add rollback support with gamecloud command rollback:

+++ resources/chef/cookbooks/deploycode/recipes/deploy.rb
@@ -25,26 +25,32 @@ execute "preparesourcefolder" do
 end
 
 script "deploycode" do
-	interpreter "bash"
-	user "root"
-	code <<-EOH
-	cd #{node[:deploycode][:localsourcefolder]}
-	export CHECK=`cat #{node[:deploycode][:localsourcefolder]}/.git/config|grep #{node[:deploycode][:gitrepo]} | wc -l`
-	if [ $CHECK -gt 0 ];then
-	git pull;
-	else
-	for x in `ls -a`
-	do
-		if [ $x != "." ] && [ $x != ".." ];
-		then
-		rm -rf $x
-		fi
-	done
-	git clone --depth 1 #{node[:deploycode][:gitrepo]} . 
-	fi
+        interpreter "bash"
+        user "root"
+        code <<-EOH
+        cd #{node[:deploycode][:localsourcefolder]}
+        export CHECK=`cat #{node[:deploycode][:localsourcefolder]}/.git/config|grep #{node[:deploycode][:gitrepo]} | wc -l`
+        if [[ #{node[:deploycode][:gitrepo]} == rollback* ]] ;
+        then
+                tag=`echo #{node[:deploycode][:gitrepo]} | cut -d':' -f 2`
+                git fetch && git checkout $tag
+                exit 0
+        fi
+        if [ $CHECK -gt 0 ];then
+        git pull;
+        else
+        for x in `ls -a`
+        do
+                if [ $x != "." ] && [ $x != ".." ];
+                then
+                rm -rf $x
+                fi
+        done
+        git clone --depth 1 #{node[:deploycode][:gitrepo]} .
+        fi
         git tag -a v_`date +"%Y%m%d%H%M%S"` -m 'Code Deploy'
         git push --tag
-	EOH
+        EOH
 end
 
 execute "changeowner" do
